organization, creating classes for each type of tab
organization, creating classes for each type of tab

moving/removing many of the files in the first commit.

created new classes in meta.py. Prior to this commit, dictionaries were being passed around between functions. easy to lose track.

hopefully this will allow for instantiation of different tabs in a single UI without dependency issues (we don't want the tab for reconstruction to depend on the tab for alignment).